requirements.txt
sys
re
bcrypt~=4.1.1
PyMySQL~=1.1.0
python-dotenv
pyotp
pyqrcode